requirements.txt
pandas
numpy
requests
scikit-learn
matplotlib
jupyter
python-dotenv
